Moved LongClickListener for Archive node
from HomeMenuActivity to new class HomeMenuJiveItemView
